Moving the batch file tests to a separate class
