+++ src/Dhhr.KppParser.Service.Tests/KppServiceTests.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.IO;
 using System.Linq;
 using Dhhr.KppParser.Service.Models;
 using Dhhr.KppParser.Service.Utils;
@@ -11,6 +10,12 @@ namespace Dhhr.KppParser.Service.Tests
     [TestClass]
     public class KppServiceTests : TestBase
     {
+        [TestInitialize]
+        public void TestInit()
+        {
+            TestInitBase();
+        }
+
         [TestMethod]
         public void KppService_ShouldSetMetadata_WhenParsingKpp()
         {
@@ -162,178 +167,6 @@ namespace Dhhr.KppParser.Service.Tests
             // assert
             isValid.Should().Be(expectedError == null);
             errorMessage.Should().Be(expectedError ?? string.Empty);
-        }
-
-        [TestMethod]
-        public void KppService_ShouldCreateBatchFiles_WhenBatchFileCreationIsEnabled_AndResultingKppFileExceedsSizeLimit_AndEpisodeFileContainsSingleInstitutionId()
-        {
-            // arrange
-            var args = BatchFileArgs("episode_institusjon.csv");
-
-            // act
-            KppService.Run(args, null, null);
-
-            // assert
-            var singleFileExists = File.Exists(OutputFile());
-            Assert.IsFalse(singleFileExists);
-
-            var batchFilesExist = OutputBatchFiles().Length > 0;
-            Assert.IsTrue(batchFilesExist);
-        }
-
-        [TestMethod]
-        public void KppService_ShouldAbortProcess_WhenBatchFileCreationIsEnabled_AndResultingKppFileExceedsSizeLimit_AndEpisodeFileContainsMultipleInstitutionIds()
-        {
-            // arrange
-            var args = BatchFileArgs();
-            args.EpisodePath = TestDataPath("episode_institusjoner.csv");
-
-            // act
-            KppService.Run(args, null, null);
-
-            // assert
-            var singleFileExists = File.Exists(OutputFile());
-            Assert.IsFalse(singleFileExists);
-
-            var batchFilesExist = OutputBatchFiles().Length > 0;
-            Assert.IsFalse(batchFilesExist);
-        }
-
-        [TestMethod]
-        public void KppService_ShouldCreateSingleFile_WhenBatchFileCreationIsEnabled_AndResultingKppFileIsWithinSizeLimit()
-        {
-            // arrange
-            var args = BatchFileArgs();
-            args.BatchFiles.MaxFileSizeInBytes = 7000;
-
-            // act
-            KppService.Run(args, null, null);
-
-            // assert
-            var singleFileExists = File.Exists(OutputFile());
-            Assert.IsTrue(singleFileExists);
-
-            var batchFilesExist = OutputBatchFiles().Length > 0;
-            Assert.IsFalse(batchFilesExist);
-        }
-
-        [TestMethod]
-        public void KppService_ShouldCreateSingleFile_WhenBatchFileCreationIsDisabled()
-        {
-            // arrange
-            var args = DefaultArgs();
-
-            // act
-            KppService.Run(args, null, null);
-
-            // assert
-            var singleFileExists = File.Exists(OutputFile());
-            Assert.IsTrue(singleFileExists);
-
-            var batchFilesExist = OutputBatchFiles().Length > 0;
-            Assert.IsFalse(batchFilesExist);
-        }
-
-        [TestMethod]
-        public void KppService_BatchFiles_ShouldAllHaveEqualLopenr()
-        {
-            // arrange
-            var args = BatchFileArgs();
-
-            // act
-            KppService.Run(args, null, null);
-
-            // assert
-            var batchLopenr = OutputBatchFiles()
-                .Select(XmlUtils.DeserializeFromFile<MsgHead>)
-                .Select(msgHead => msgHead.Items.Single().As<Document>())
-                .Select(doc => doc.RefDoc.Item.As<RefDocContent>())
-                .Select(content => content.Melding.lopenr)
-                .ToList();
-
-            batchLopenr.Should().OnlyContain(lopenr => lopenr == batchLopenr[0]);
-        }
-
-        [TestMethod]
-        public void KppService_BatchFiles_ShouldAllConformToSpecificLokalidentFormat()
-        {
-            // arrange
-            var args = BatchFileArgs();
-
-            // act
-            KppService.Run(args, null, null);
-
-            // assert
-            var messages = OutputBatchFiles()
-                .Select(XmlUtils.DeserializeFromFile<MsgHead>)
-                .Select(msgHead => msgHead.Items.Single().As<Document>())
-                .Select(doc => doc.RefDoc.Item.As<RefDocContent>())
-                .Select(content => content.Melding);
-
-            foreach (var message in messages)
-            {
-                var lokalidentParts = message.lokalident.Split('_');
-
-                lokalidentParts[0].Should().Be(message.lopenr);
-                lokalidentParts[1][0].Should().Be('(');
-                lokalidentParts[2].Should().Be("Of");
-                lokalidentParts[3][^1].Should().Be(')');
-
-                var valueBeforeOf = lokalidentParts[1][1..];
-                var valueAfterOf = lokalidentParts[3][..^1];
-
-                int.TryParse(valueBeforeOf, out var batchMessageNumber).Should().BeTrue();
-                int.TryParse(valueAfterOf, out var messagesInBatch).Should().BeTrue();
-
-                batchMessageNumber.Should().BeLessThanOrEqualTo(messagesInBatch);
-            }
-        }
-
-        [TestMethod]
-        public void KppService_BatchFiles_ShouldContainTheSameEpisodes_AsSingleFile_WhenSameEpisodeFileIsUsedAsInput()
-        {
-            // arrange
-            const string episodeFileName = "episode_institusjon.csv";
-
-            var batchFileArgs = BatchFileArgs(episodeFileName);
-            var singleFileArgs = DefaultArgs();
-
-            singleFileArgs.EpisodePath = TestDataPath(episodeFileName);
-
-            // act
-            KppService.Run(batchFileArgs, null, null);
-            KppService.Run(singleFileArgs, null, null);
-
-            // assert
-            var batchMessages = OutputBatchFiles()
-                .Select(XmlUtils.DeserializeFromFile<MsgHead>)
-                .Select(msgHead => msgHead.Items.Single().As<Document>())
-                .Select(doc => doc.RefDoc.Item.As<RefDocContent>())
-                .Select(content => content.Melding);
-
-            var singleMessage = XmlUtils.DeserializeFromFile<MsgHead>(OutputFile())
-                .Items.Single().As<Document>()
-                .RefDoc.Item.As<RefDocContent>()
-                .Melding;
-
-            var episodesInBatchMessages = batchMessages
-                .SelectMany(message => message.Institusjon.Single().Objektholder.Single().EpisodeKPP)
-                .ToArray();
-
-            var episodesInSingleMessage = singleMessage.Institusjon.Single().Objektholder.Single().EpisodeKPP;
-
-            episodesInBatchMessages.Should().BeEquivalentTo(episodesInSingleMessage);
-        }
-
-        private Args BatchFileArgs(string episodeFileName = "episode_institusjon.csv")
-        {
-            var args = DefaultArgs();
-
-            args.BatchFiles.EnableCreation = true;
-            args.BatchFiles.MaxFileSizeInBytes = 3000;
-            args.EpisodePath = TestDataPath(episodeFileName);
-
-            return args;
         }
     }
 }
